Initial commit
Core of how the API will be based on. Right now the user can open
multiple browsers (screens) and it will see a cube moving through
all the screens, and the user can add and remove screens as many
screens as they want. It has a really simple server and clients
logic. Every time a cube reaches the end of the screen the server
sends a notification to all clients (sockets) saying which screen
is next, and then the socket that matches the id can show the same
animation.