Adding all cases of browser closure
In this code we specify 3 possible actions that the system might taking when closing a socket. Mirroring the browsers so that they will be added to the left instead of the right. Moving one, when a socket we close and the system simply swtches one closer to the center. And reorganising which is when one side has way more sockets that the other, in which case the center needs to change.