Adding feature for moving the camera
With this new code the user can move the angle of the camera by pressing the arrow keys. This will only translate the camera (and therefore the objects). This was achieved by simply sending a message to all the sockets to translate the x,y,z values of the camera accordingly.