Adding moving user features
When pressing keys w and s, the camera will move forwards and backwards; when pressing the keys a and d the camera will move left and right; when pressing the up and down arros keys the camera will move up and down. When dragging the mouse vertically the camera will rotate around the X axis, when dragging the mouse horizontally the camera will rotate around the Y axis; and when pressing the q and e keys the camera will rotate around the Z axis. Pressing the r key will reset the view.