Rotation animation + dimensions of window
I have updated the method for rotating of an object, for issue #11. This has a similar approach for the translation. In addition, I did a study of what is the good angle for rotation for wide windows and narrow windows to quickly being able to assign the the window size and rotate the neightbour windows accordingly, for issue #7